separated moviecart component
