+++ Services/index.ts
@@ -1,0 +1,61 @@
+export const getGenresByID = id => {
+    if (id == "28") {
+        return "Action";
+    }
+    if (id == "12") {
+        return "Adventure";
+    }
+    if (id == "16") {
+        return "Animation";
+    }
+    if (id == "35") {
+        return "Comedy";
+    }
+    if (id == "80") {
+        return "Crime";
+    }
+    if (id == "99") {
+        return "Documentary";
+    }
+    if (id == "18") {
+        return "Drama";
+    }
+    if (id == "10751") {
+        return "Family";
+    }
+    if (id == "14") {
+        return "Fantasy";
+    }
+    if (id == "36") {
+        return "History";
+    }
+    if (id == "27") {
+        return "Horror";
+    }
+    if (id == "10402") {
+        return "Music";
+    }
+    if (id == "9648") {
+        return "Mystery";
+    }
+    if (id == "10749") {
+        return "Romance";
+    }
+    if (id == "878") {
+        return "Science Fiction";
+    }
+    if (id == "10770") {
+        return "TV Movie";
+    }
+    if (id == "53") {
+        return "Thriller";
+    }
+    if (id == "10752") {
+        return "War";
+    }
+    if (id == "37") {
+        return "Western";
+    } else {
+        return "Other";
+    }
+};
